refactor(core): extract domain types
Move the plain-data DTOs that describe what every WaveFlow client
sees (`Track`, `TrackListItem`, `ListTracksResponse`, `Profile`,
`Library`, `Playlist`, plus their `Create*Input` / `Update*Input`
counterparts) out of `commands/{track,profile,library,playlist}.rs`
and into `crates/core/src/domain/`.

The original command modules keep a `pub use waveflow_core::domain::…`
re-export so existing call sites (`crate::commands::profile::Profile`,
the `sqlx::query_as::<_, Profile>(...)` patterns, etc.) keep working
without churn. The row-shaped types still tied to specific SQL
projections (`TrackRow`, the private query targets in
`commands/playlist.rs`) stay in app — they migrate naturally with
the repository traits in step 5.

`waveflow-core` picks up a single optional dep (`sqlx`) gated behind
a new `sqlite` feature. Each `FromRow` derive on the moved types is
`#[cfg_attr(feature = "sqlite", derive(sqlx::FromRow))]`, so the
crate stays storage-agnostic when the feature is off (the future
`waveflow-server` will toggle a `postgres` equivalent without
double-declaring the structs). Desktop opts in via
`waveflow-core = { path = "../core", features = ["sqlite"] }`.

`cargo check --workspace --all-targets` and `cargo test --workspace`
(111 passed) both green. Net diff: 225 LOC removed from app, 45 LOC
added (re-exports + the feature flag wiring).

Refs #128.
